requirements.txt
fastapi
uvicorn
pymupdf
faiss-cpu
sentence-transformers
requests
python-multipart
